Improve module augmentation in `page-break` plugin
