+++ packages/ckeditor5-page-break/src/index.ts
@@ -10,3 +10,7 @@
 export { default as PageBreak } from './pagebreak';
 export { default as PageBreakEditing } from './pagebreakediting';
 export { default as PageBreakUI } from './pagebreakui';
+
+export type { default as PageBreakCommand } from './pagebreakcommand';
+
+import './augmentation';
