+++ packages/ckeditor5-page-break/src/pagebreakcommand.ts
@@ -72,9 +72,3 @@ function getInsertPageBreakParent( selection: DocumentSelection, model: Model ):
 
 	return parent as Element;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		pageBreak: PageBreakCommand;
-	}
-}
